+++ hl_image_picker/example/lib/main.dart
@@ -66,6 +66,7 @@ class _MyHomePageState extends State<MyHomePage> {
           maxSelectedAssets: _isCroppingEnabled ? 1 : _count,
           usedCameraButton: _usedCameraButton,
           numberOfColumn: _numberOfColumn,
+          isGif: true,
         ),
         cropOptions: HLCropOptions(
           aspectRatio: _aspectRatio,