+++ hl_image_picker_ios/lib/hl_image_picker_ios.dart
@@ -99,6 +99,7 @@ class HLImagePickerIOS extends HLImagePickerPlatform {
       'minFileSize': pickerOptions?.minFileSize,
       'usedCameraButton': pickerOptions?.usedCameraButton,
       'numberOfColumn': pickerOptions?.numberOfColumn,
+      'isGif': pickerOptions?.isGif,
       'cropMaxWidth': cropWidth,
       'cropMaxHeight': cropHeight,
       'localized': localized?.toMap(),